Update touch.py
Added a try, except: around f.read(25). This ensures that if it cant read from the device then it will exit the function then reconnect to the device. In my case whenever i plugged in any USB it would disconnect and change the driver name. this should fix that problem.